+++ src/commonMain/kotlin/events/Events.kt
@@ -3,9 +3,9 @@ package events
 import Weapon
 import com.soywiz.korge.view.View
 
-data class BulletHitEvent(val target: View, val damage: Double)
+data class BulletHitEvent(val target: View, val damage: Int)
 
-data class EnemyAttackEvent(val damage: Double)
+data class EnemyAttackEvent(val damage: Int)
 
 object EnemyDiedEvent
